examples/task-manager-bot/pyproject.toml
[project]
name = "task-manager-bot"
version = "1.0.0"
description = "A personal task manager Telegram bot built with Botty framework"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "botty>=0.0.1",
    "python-telegram-bot>=21.0",
    "sqlmodel>=0.0.14",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
]


[tool.uv.sources]
botty = { workspace = true }
